Staging

+++ route/route-photo.js
@@ -27,14 +27,13 @@ module.exports = function(router) {
 
     .post(bearer_auth_middleware, bodyParser, (req, res) => {
       return upload(req, res, err => { 
-        debug('err', err);
         if(err) return errorHandler(err, res);
         return multiPart(req, res);
       });
       function multiPart(req, res) {
+        if (!req.body.name || !req.body.description ||  !req.body.gallery_id) return errorHandler(new Error('Validation Error:  Required fields missing.'),res);
         Photo.upload(req)
           .then(data => {
-            debug('data', data);
             return new Photo(data).save();
           })
           .then(img => res.status(201).json(img))
@@ -48,10 +47,8 @@ module.exports = function(router) {
           .then(img => res.status(200).json(img))
           .catch(err => errorHandler(err,res));          
       }
-      debug('req.user.user_id', req.user._id);
       return Photo.find({user_id: req.user._id})
         .then(imgs => {
-          debug('imgs', imgs);
           return imgs.map(img => img._id);
         })
         .then(img_ids => res.status(200).json(img_ids))
@@ -59,9 +56,9 @@ module.exports = function(router) {
     })
 
     .delete(bearer_auth_middleware, (req, res) => {
-      debug('req.params.id', req.params.id);
       return Photo.findById(req.params.id) //find one
         .then(img => {
+          if (!img) return new Error('Error ENOENT: Resource does not exist');
           if (img.user_id.toString() !== req.user._id.toString()) return new Error('Authorization Error: permission denied');
           return img.delete();
         })
